fix: add missing EditorSettings methods and fix autoSaveInterval type
- EditorSettings: add syntaxHighlighting field + withFontSize, withTabSize, withShowLineNumbers, withWordWrap, withSyntaxHighlighting methods
- AppSettings: change autoSaveInterval type Int→Long to match ViewModel usage

+++ app/src/main/java/com/mobilecodex/model/AppSettings.kt
@@ -10,7 +10,7 @@ data class AppSettings(
     val enableAnimations: Boolean = true,
     val compactMode: Boolean = false,
     val autoSave: Boolean = true,
-    val autoSaveInterval: Int = 300,
+    val autoSaveInterval: Long = 300,
     val showNotifications: Boolean = true,
     val hapticFeedback: Boolean = true,
     val editorSettings: EditorSettings = EditorSettings()
@@ -21,7 +21,7 @@ data class AppSettings(
     fun withEnableAnimations(enable: Boolean) = copy(enableAnimations = enable)
     fun withCompactMode(compact: Boolean) = copy(compactMode = compact)
     fun withAutoSave(enabled: Boolean) = copy(autoSave = enabled)
-    fun withAutoSaveInterval(interval: Int) = copy(autoSaveInterval = interval)
+    fun withAutoSaveInterval(interval: Long) = copy(autoSaveInterval = interval)
     fun withShowNotifications(enabled: Boolean) = copy(showNotifications = enabled)
     fun withHapticFeedback(enabled: Boolean) = copy(hapticFeedback = enabled)
